requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2==2.9.9
PyJWT==2.9.0
requests==2.32.3
SQLAlchemy==2.0.32
twilio==9.2.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
gunicorn==20.1.0
